+++ edbee-lib/edbee/texteditorwidget.cpp
@@ -53,7 +53,7 @@ TextEditorWidget::TextEditorWidget( QWidget* parent)
     , scrollAreaRef_(0)
     , editCompRef_(0)
     , autoCompleteCompRef_(0)
-	, autoScrollMargin_(50)
+    , autoScrollMargin_(50)
 {
     // auto initialize edbee if this hasn't been done alread
     Edbee::instance()->autoInit();
@@ -251,6 +251,18 @@ void TextEditorWidget::setHorizontalScrollBar(QScrollBar* scrollBar)
 {
     scrollAreaRef_->setHorizontalScrollBar(scrollBar);
     emit verticalScrollBarChanged( scrollBar );
+}
+
+/// Returns the auto scroll margin
+int TextEditorWidget::autoScrollMargin() const
+{
+    return autoScrollMargin_;
+}
+
+/// Sets the auto scrollmargin
+void TextEditorWidget::setAutoScrollMargin(int amount)
+{
+    autoScrollMargin_ = amount;
 }