+++ edbee-lib/edbee/views/components/texteditorautocompletecomponent.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "edbee/exports.h"
+
 #include <QModelIndex>
 #include <QWidget>
 #include <QLabel>
@@ -46,7 +48,7 @@ protected:
 
 /// An autocomplete list
 /// Which receives it's autocomplete list from the document
-class TextEditorAutoCompleteComponent : public QWidget
+class EDBEE_EXPORT TextEditorAutoCompleteComponent : public QWidget
 {
     Q_OBJECT
 public: